WL#8817 Dbacc::ElementHeader zero marks no element.
An aesthetic refactoring making an ElementHeader with only zero bits usable as
a placeholder for later valid ElementHeader value.

Before this patch a zero value was interpreted as a locked element using
operation record #0.

Now it could be misread as an unlocked, and unscanned element with an invalid
reduced hash value (which must always contain at least one set bit).