Bug #20933307: VIEWS CAN EVALUATE USER FUNCTIONS BEFORE                VIEW RESTRICTIONS
ISSUE:
------
VIEWs are a common means to implement row-level filtering.
Such VIEWs have filter-conditions on which rows are
accessible to end users.  However, in certain conditions,
the optimizer may choose to defer evaluating the
restrictions contained in the VIEW definition if they are
determined to be more expensive than other, user-supplied
conditions.

This is problematic when user-supplied conditions include
UDFs or SQL functions which can record row data before VIEW
filters have been applied.

SOLUTION:
---------
While merging the view's underlying table with the query,
we swap the evaluation order to bring the predicates from
the view definition ahead of the outer-query's predicate.

This change happens only when there is a stored procedure
in the outer-query's predicate.