Bug #22647379:  FAILURE TO BUILD NDB UNIT TESTS DUE TO MYTAP LINK FAILURE
The unit tests for NDB were built regardless of the WITH_UNIT_TESTS
switch. After commit 084a9547b7e69188d74b1207d673dc00e0e884c5
for Bug #22246885, the tests depend on the mytap library which is
only built if the MySQL unittests are built. Thus, a link failure
occurs when the build is performed without unit tests.

A check for the WITH_UNIT_TESTS switch was included in the
NDB_ADD_TEST macro. Now the NDB unittests are built only
when requested.