Updating the ndb_addnode_restart test cases.
The ndb_addnode_restart is run on only one table and only once since
after that first run all the nodes would be already added to the
cluster and the case - addnode and restart cannot be tested. But the
step runAlterTableAndOptimize function is written to handle multiple
tables in a single run.

This patch removes the unnecessary loop and modifies the
runAlterTableOptimize to handle only a single table.

(followup to commit 1038d067b)