Updating the ndb_addnode_restart test cases.
The re-organize table step sometimes runs and completes even before the
other step could kill the nodes. This would be reported as a failure by
the autotest. To ensure that the re-organize waits until the other step
is about to start killing the nodes, this patch intoduces a variable -
WaitForNodeKillStart to the test case. Now the 'alter reorganize' step
waits until the other step starts killing the nodes by checking this
new variable.

(followup to commit 1038d067b)