WL#6378 New Data Dictionary
Problem:

The class dd::Base_collection is a interface which enables
storing collection of DD objects. This collection is used to
store DD objects like dd::Column, dd::Index, dd::Tablespace_file,
dd::Trigger and many more.

The objects in dd::Base_collection are all created at once and
dropped at once. There is also a possibility that a single object
from the collection is removed. The problem in hand is that
attempt to update [dd::cache::Dictionary_client::update()] the
DD object which own this collection fail to remove collection
item from DD tables. The following sequence has the problem:

1. Create a DD object with single item in collection and
store() it.
2. Acquire and drop the item from collection.
3. Update [dd::cache::Dictionary_client::update()] the DD object.

Operation 3. fails to remove item from DD table.

These operations are not being tested now as addition, removal of
columns and indexes will endup creating a whole new dd::Table and
it does not modify existing dd::Table object. We stumbled on this
issue when work on wl7896 testing to drop a trigger.

Solution:
The problem was that dd::Base_collection::is_empty() did only
check if there exists items (m_items) in the collection, but
ignored items that are requested to be removed (m_removed_items).
The fix is to say that the collection is empty only if both of
these m_items and m_removed_items are empty in a collection.