Bug#21953132 RELAY LOG INFO CURRENTLY_EXECUTING_GTID IS NOT PROPERLY              INITIALIZED OR PROTECTED
Problem:

When replication is configured but not started on a slave, the relay
log info variable "currently_executing_gtid" is not properly
initialized and would be used if performance_schema table
replication_applier_status_by_worker be queried.

A valgrind test confirmed that the variable can be used uninitialized.

Fix:

Initialized the variable properly on relay log info constructor.