Bug #22281205: CRASH IN FIELD_NUM::PREPEND_ZEROS
Issue:
-----
When a numeric value is converted to a zerofill string,
null value is not handled correctly.

SOLUTION:
---------
After calling val_str, checking for null should be done
using null_value, not is_null(). As the comment in
Item::is_null() suggests, it should be used only to
determine the nullness of a complex expression without
fully evaluating it.