Bug#22615274: API CRASHES WITH SIGNAL 6 AT TRANSPORTERFACADE.CPP LINE 2170
Root cause for this bug was a misconseption of how
TransporterFacade::do_poll() worked: The sequence of
'poll' calls needed related to do_poll() are:

- ::start_poll()
- 'construct signal' -> raw_sendSignal()
- 'do_forceSend()
- **do_poll()**
- complete_poll()

That has led us to belive that do_poll() would force *this* thread
to grab the poll-right and perform the transporter poll, whicheventually should
end up in ::trp_deliver_signal() calling
TransporterFacade::expand_clnt().

However, as it turned out, all do_poll() does for us is
to ensure that *some thread* will get the poll-right. That
could be this thread if no other had the poll-right, or it
could be someone else already holding it.

That ripped away the foundation for how the m_open_close_mutex
was used to protect the TransporterFacade::m_threads members.
We used to grab that lock early in TransporterFacade::open_clnt(),
and held it across do_poll() while waiting for the EXPAND_CLNT
signal to be executed. However, if another thread had the
poll right and executed that signal for us, that broke down
the entire open_close_mutex protection 'protocol'.

This patch temporary unlocks the m_open_close_mutex while ::open_clnt
wait for the EXPAND_CLNT signal to be executed. The thread
executing that signal will then grab the lock instead in
TransporterFacade::expand_clnt().

A similar defect in the m_open_close_mutex protection was
also identified in TransporterFacade::close_clnt().
That is also fixed by this patch.