Bug#22283790: RANGE OPTIMIZER UTILIZES TOO MUCH MEMORY WITH MANY OR CONDITIONS
Problem:
The default value for range_optimizer_max_mem_size exceeds if
too many OR conditions are there.

Analysis:
For every OR condition, range optimizer creates a SEL_TREE object. Although
it keeps merging the SEL_TREE objects if it can be done, it cannot delete the
SEL_TREE objects memory.As a result, when there are many OR conditions,
we see that the default value of range_optimizer_max_mem_size exceed
very quickly.
range_optimizer_max_mem_size is actually given that default keeping a SEL_ARG
object's memory in mind. Memory take for a SEL_ARG object is 105 bytes, while
the memory taken for a SEL_TREE object is close to 600 bytes. Hence the memory allocated
gets exhausted very easily when too many SEL_TREE's are created.

Solution:
A SEL_TREE object has an array of 64 SEL_ARG pointers. This is based on the fact
that a table can have a max of 64 Keys. However, it is not needed for every
SEL_TREE object to have those many pointers. RANGE_OPT_PARAM has a member called
"keys" which defines the number of keys that are considered for optimization.
Using this we make SEL_ARG **keys[MAX_KEY] to Mem_root_array<SEL_ARG *,true> keys;
And reserve the memory required for the array using RANGE_OPT_PARAM's "keys" member.

With this improvement we have reduced the memory usage down for range optimizer.
As a result, the new default for range_optimizer_max_mem_size is 8388608 bytes.