BUG#22650195 - ASSERT IN ROW_PARSE_INT()
The updated counter would be got from the update vector after the
differences are calculated. The issue is due to we check the counter
field by checking the field_no in upd_field_t, which could be the
field number in either clustered index or secondary index(including
virtual index). So if it's a virtual index, we could get the incorrect
field for checking, thus the assertion failure.

The solution is that we should check the field_no along with the field
of the virtual column, if it was a virtual column on virtual index,
we should ignore this field.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 11795