BUG#22699395 - SQL_THREAD OF BINLOG-LESS SLAVE FAILS WITHOUT ANY MESSAGE FOR VIEW_CHANGE EVENT
Problem :

The problem with this bug is that when the binlogless slave is
trying to read from an active member of the group, the view_change event
is not applied on the slave and therefore it errors out. The error
happens in the view_change_event::do_apply_event when the view_change
event is being written to the binary log which is disables as the slave
is started with skip-log-bin option ON.

Solution :

The problem is fixed by skipping applying the event for the case the
binary log is disabled on the slave which is replicating from the group.
This is achieved by checking the binlog server option.