Bug #22511846 NDBT FRAMEWORK CAN LEAK NDB OBJECTS
~ndb_cluster_connection now waits for all subtended Ndb objects
to be deleted.

  Bug  19999242 DELETEING NDB_CLUSTER_CONNECTION WITH NDB INSTANCES

This requires that NdbApi applications delete their Ndb objects
before the ndb_cluster_connection.
Some applications and test programs based on the NDBT framework
have been observed not to do this, particularly on exit
paths.

This patch adds cleanup of NDBT allocated Ndb objects on error
paths.