Bug#22734050 : MAIN.RANGE_ALL AND ALL THE RANGE TESTS FAIL ON SOLARIS
Problem:
The test for checking if range optimizer bails out when too many SEL_ARG's
were created takes a long time now to complete. This is within the timeframe
set by mtr for some machines and not so for slower machines.

Analysis:
Before the fix for Bug#17769777, the test in range.inc used to pass because
there was a check for the MAX_SEL_ARGS (16000) before calling test_use_count
in a debug build. Reason being it takes too much time to traverse a big range tree.
This was removed when fix for Bug#17769777 was made. For
the default value set for range_optimizer_max_mem_size, the test used to pass
because it exceeds the default value for less number of SEL_ARGS.
However with the fix for Bug#22283790, the default value is now raised from 1.5MB
to 8MB. This results in creating more SEL_ARGS than earlier which in turn creates
a large range tree. Hence the failure.

Solution:
Introduce the check back. The value for checking is now based on the amount of memory,
not on the number of SEL_ARGS. Value is set to 2MB because it would ideally take
~1.6MB for 16000 SEL_ARGS. For an OR condition we will be taking double the
amount of memory for the same SEL_ARGS. But if we set the value to 4MB for the check,
in case of AND conditions, it might create 35000 SEL_ARGS. This might again,
take time to traverse.