Bug#22537196 OPTIMIZER: SOURCE AND DESTINATION OVERLAP IN MEMCPY IN UPDATE_REF_AND_KEYS
Problem: The compiler may use memcpy() when generating the assignment operator
for a class/struct which does not provide its own.
Such an implementation will not check for self assignment, and is
undefined behaviour if <source> == <destination>

Solution: re-introduce the check for self-assignment introduced by the patch for
Bug#56138 "valgrind errors about overlapping memory when double-assigning same variable"

That patch assumed that the bug was fixed in gcc 4.4 and up.
Turns out it is not.