BUG#21877062: MIN/MAX IN VIEW ON TIMESTAMDIFF IN VIEW CONFUSES 	      OPTIMIZER TO THROW SYNTAX ERROR
ANALYSIS:
=========
CREATE VIEW operation using timestampdiff() function with
MICROSECOND as the unit results in an incorrect view
definition.

'Item_func_timestamp_diff::print()' is used to generate view
definition when it involves a timestampdiff operation. The
function generated an incorrect view definition since it
translated the 'MICROSECOND' unit to an invalid string
'SECOND_FRAC'. Operations on such an invalid view resulted
in a syntax error.

FIX:
====
The fix involves appending the string 'MICROSECOND' while
generating the view definition when the unit specified is
'MICROSECOND'.