Bug#21855705 MYSQLDUMP --DUMP-SLAVE DOES NOT WORK WITH MULTI-SOURCE
Problem:

Even if multi-source replication is enabled, mysqldump --dump-slave outputs
only a single CHANGE MASTER statement, setting the positions only for the
default channel.

Analysis:

mysqldump obtains the slave configuration by issuing a SHOW SLAVE STATUS
statement and collecting the output. It is collecting only the output of
the first row returned by the SHOW SLAVE STATUS, discarding all the
additional channels if they are configured.

Fix:

Instead of reading only the content of the first row of the result of the
SHOW SLAVE STATUS statement, mysqldump now keeps reading the content of
the result to also dump the CHANGE MASTER statement for the non-default
channels that are configured.