BUG#22452992 - PB2 INNODB_UNDO.TRUNCATE_RECOVER TEST FAILS
From the error log, the server refused to start because UNDO tablespaces
were not found. In fact, server should be allowed to recreate the UNDO
tablespaces and start.

The fix is to keep silence in recv_init_crash_recovery_spaces() if the
UNDO tablespaces were not found, just print the error message once.
This patch also refactors the code a bit to make it more readable.

Reviewed-by: Marko Makela <marko.makela@oracle.com>
RB: 11837