BUG#22734285: LIMIT/REMOVE THE APPLIER RECEIVED GTID SET
On Group Replication we have two types of transactions:
  local:  committed by the local member;
  remote: committed by other group members.
The local transactions follow the same flow like if this server
was a single master on a asynchronous replication.
The remote transactions are delivered by the communication layer and
queued for conflict detection and apply. These transactions, after
being accepted to commit (no conflicts detected) are queued to a
relay log and eventually applied. These transactions are tracked on
group_replication_applier channel. DBA can check them by consulting
column received_transaction_set on
performance_schema.replication_connection_status table.
Since it only contains remote transactions, if this member does
commit transactions there will be gaps, example:
  Three members group: M1, M2 and M3.
  M1 does commit UUID:1 and UUID:3
  M2 does commit UUID:2
  M2 received_transaction_set will be: UUID:1,3. UUID:2 will not be
there.
On the long term, the number of gaps may increase, and this may have
impact on the performance.

To solve the above issue, periodically, we update the received set
with the full set of transactions committed on the group, closing
the gaps.
This update is done on the same moment that all members do exchange
their committed transactions set and garbage collect the
certification info.

Changes
=======
To implement the solution for this bug, the following changes were
required:

Server side
-----------
 1. Add functionality to copy GTID_EXECUTED to replication channel
    received transaction set.
    Files: rpl_channel_service_interface.h,
           rpl_channel_service_interface.cc, rpl_rli.h, rpl_rli.cc

Validation and testing is done on BUG 22671747, on Group Replication
plugin.