Bug #21783798 NO MYSQL_AUDIT_GLOBAL_VARIABLE_GET EVENTS WHEN   SHOW_COMPATIBILITY_56=OFF
When show_compatibility_56 is off performance_schema converts
"SHOW GLOBAL VARIABLES LIKE 'var_name'" into a SELECT statement
over the performance_schema table. And since performance_schema
does always produce the full content of the table and allows the
optimizer to filter it out we have no other choice but produce
MYSQL_AUDIT_GLOBAL_VARIABLE_GET events for all of the global
variables.
Note that this is different from the case when show_compatibility_56
is off since the INFORMATION_SCHEMA does pass down the condition on
the I_S tables and the other variables are not queried.

Updated the test case to test SHOW GLOBAL VARIABLES LIKE 'gizmo' for
both on and off values of show_compatibility_56.