Bug#22811150 SQLCOM_EMPTY_QUERY should be considered GTID innocent
Problem:

A statement that was skipped because of a "magic" version comment is
being considered as able to change the database/database data and so
it is not considered as "innocent" from the GTID point of view.

Analysis:

Some statements (like SET statements for example) are considered
"innocent" because they will not change things in the database or in
the database data.

It is always possible to set GTID_NEXT session variable when only
"innocent" statements were executed.

A SQLCOM_EMPTY_QUERY statement is not being considered as innocent
from the GTID execution point of view.

Fix:

Added SQLCOM_EMPTY_QUERY to the list of innocent statements for GTIDs.