Bug#22816576: ASSERTION FAILED: !THD->IS_ERROR() IN OPTIMIZE_COND() [II]
Problem: The functions that parse JSON path arguments don't
distinguish between a NULL path and a syntactically invalid path. This
upsets an assert statement that checks that the returned error status
is consistent with the session's error flag, and causes failures in
debug builds.

Solution: Change the parsing of JSON path arguments so that it clearly
distinguishes between valid non-NULL paths, NULL paths and invalid
paths.