Bug#22782197 ACCESSING UNINITIALIZED MEMORY INSIDE MYSYS/MY_MALLOC.C MY_FREE() LINE ~129
Before this fix, when compiling the server with tools like Valgrind,
the tool could report that calls to memory_free_noop()
 use uninitialized memory for the last parameter (struct PSI_thread * owner)

The report is technically correct, but the issue has no consequence,
as the parameter (even with a random value) is not used.

With this fix,
  - memory_alloc_noop()
  - memory_realloc_noop()
now cleanly return a properly initialized NULL value
for the output 'owner' parameter,
which is what is eventually passed to memory_free_noop().