Bug#22227958: ASSERT IN DD::FILL_DD_TABLE_FROM_CREATE_INFO, DDL
Problem: The new DD code had added an assert which checked that the bitmask
flags for HA_OPTION_STATS_PERSISTENT and HA_OPTION_NO_STATS_PERSISTENT could
not be set at the same time. This is reaonable, but unfortunatly it is
legal to supply both values, both when creating the table and when altering
it. The same problem affects HA_OPTION_PACK_KEYS/HA_OPTION_NO_PACK_KEYS,
HA_OPTION_CHECKSUM/HA_OPTION_NO_CHECKSUM, and HA_OPTION_DELAY_KEY_WRITE
/HA_OPTION_NO_DELAY_KEY_WRITE.

Solution: Change the parser so that it always clears both flags before
setting any of these flags. That way there will be no visible change in behavior
and only the last value set will be reflected in the bitfield.