Bug #22811659 "ACCESSES TO NEW DATA-DICTIONARY ADD CONFUSING STAGES TO P_S.EVENTS_STAGES_*.".
With advent of new data-dictionary DML and DDL statements started to
open and lock data-dictionary tables in order to read and/or update
table meta-data.

These operations on system tables are reflected in the same way as
opening and locking of user tables in P_S.EVENTS_STAGES_* tables and
in I_S.PROCESSLIST which can be pretty confusing.

This patch solves this problem by introducing two new stages
"Opening system tables" and "Locking system tables" to be used
instead of "Opening tables" and "System lock" stages in cases when
we access data-dictionary and other system tables (e.g. time zone,
replication or log tables).