Bug#21178196 : UPDATE MAY USE INDEX MERGE WITHOUT ANY REASON 		  (INCREASING CHANCES FOR DEADLOCK)
Problem:
Index merge access plan is choosen over range scan when the cost
for doing range scan would be same/less.

Analysis:
W.r.t the query presented in the bugpage, index merge access plan
is choosen because it is covering. However, while calculating the cost
for doing index merge scan, the current cost model thinks doing a sequential
scan is more costly than doing a random scan. For the given case,
while calculating the cost for disk sweep, we calculate cost for
reading the blocks having the qualifying rows and the cost for skipping
the other blocks. There is a minimal cost for skipping blocks too. And given
that in this case, we are looking at one qualifying row and rejecting 1590000
rows, we would end up skipping a lot of blocks. As a result the total cost
for doing sequential scan becomes larger than doing a random scan.

This in turn makes optimizer think doing index merge is better than the
other access plans.

Solution:
If it is calculated that sequential scan cost is more than random scan cost,
set the cost to random scan cost.

Also, set is_interrupted to true before calling get_sweep_read_cost
when the plan is not single table. This will make both single table select
and single table update choose the same code path. This is in line with
other calls made to get_sweep_read_cost.

Test case is not added as large amount of data is needed to get a repeatable
scenario.