Bug #21128318: NDBREQUIRE() SHOULD REPORT ACTUAL CONDITION EXPRESSION LIKE ASSERT() DOES
ndbrequire() evaluates a given expression and terminates the process if
the expression doesn't evaluate to true but does not preserve the failing
expression. The relevant error object field in the error log is logged as
follows:

<snip>
Error object: <block name> (Line: <line number>) <config variables' status>

Updated the working of ndbrequire() to preserve the expression and
include it in the error logs. The updated error object field in the error
log is now as follows:

<snip>
Error object: <block name> (Line: <line number>) <config variables' status>
               Check <failed expression> failed