Bug #21220766: MYSQLPUMP DOES NOT FILTER OUT OBJECT NAMES CORRECTLY WHEN HAVING '.' OR '@'
Added a 2 pass parser for the include and exclude options that support escaping and quoting.
It does one pass to unquote/escape the comma separated list (entries can be quoted with
double quotes, can be escaped with a backslash, the comma is a separator and non-quoted
whilespace is trimmed from both ends). And it does a second pass on each individual
option (entries can be quoted with double quotes, can be escaped with a backslash, the at
sign (@) is used as a delimiter and non-quoted whitespace is trimmed from both ends).

Added unit tests to verify the functionality.