Bug#21371070 [ERROR] INNODB: CANNOT ALLOCATE 0 BYTES: SUCCESS
Issue:
======
If the buffer pool file is empty and if we do a buffer pool load, then we
try to allocate memory to store dump info present in the dump file which
results in the "INNODB: CANNOT ALLOCATE 0 BYTES: SUCCESS"  error.

Fix:
====
Avoid allocating memory if the buffer pool file is empty.

RB: 11913
Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>