Bug#16576959: MYSQL_OPTIONS4: POTENTIAL DOS VIA PRINTING TOO MUCH USELESS INFO TO ERROR LOG
The client library can enrich connection data with information about
the library version, the client's name, the OS the client is running
on, and in fact pretty much anything it wants.  The designated
mechanism for this is mysql_options4(), which lets the client put
arbitrary key/value pairs on the connection info before the connection
is made.

The server outright refuses a connection when 64+ KB of these data are
submitted. However, in practice performance schema usually reserve much
less space (< 1 KB) for each connection's attributes. If the client submits
more data, the connection will succeed, but a warning will be thrown
that some attributes were discarded. (The server will provide access
to all attributes that is has a complete copy of. We do not currently
show how many attributes were lost on a given connection, nor do we
show an incomplete attribute with some sort of "cut" marker -- we simply
don't show a truncated attribute at all. That said, we do expose info
about the total number of connections with dropped attributes.)

Pre-patch, we simply logged a warning indicating that attributes were
truncated on a connection, without giving any information beyond the
timestamp that would allow the administrator to identify the client,
user, or connection that lost attributes. This made debugging harder
than it needs to be.

This patch adds additional data to log entry where available, such
as user/host, connection ID, etc., all in the hope that it will
make debugging such instances with just the error log much easier,
and aid cross-referencing in the presence of further logs, such as
the general query log.

It also adds a new global PFS variable,
"Performance_schema_session_connect_attrs_longest_seen" that shows the
size of the longest (valid, i.e. <= 64 KB) buffer we were passed, so
the DBA may adjust the buffer, or complain to the application developer.

Finally, when a valid buffer in excess of what we accept is passed,
we will add a new attribute "_truncated" stating how many characters
were lost (provided the configured buffer is large enough to hold
this information). This should make it easier to identify the exact
connection that had the questionable attribute set in the PFS,
without having to resort to the error log.